@@ -1,27 +1,26 @@
+import static Domain.Reservation.usedCapacity;
+import static org.junit.Assert.assertEquals;
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import com.pholser.junit.quickcheck.Property;
+import com.pholser.junit.quickcheck.runner.JUnitQuickcheck;
+
+import org.junit.runner.RunWith;
+
 import Domain.Reservation;
 import InterfaceAdapters.KVSInMemoryImpl;
 import InterfaceAdapters.TraceConsoleImpl;
 import UseCases.KeyValueStore;
 import UseCases.ReservationUseCase;
 import UseCases.Trace;
-import com.pholser.junit.quickcheck.Property;
-import com.pholser.junit.quickcheck.runner.JUnitQuickcheck;
-import org.javatuples.Pair;
-import org.junit.runner.RunWith;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-import static Domain.Reservation.usedCapacity;
-import static org.junit.Assert.assertEquals;
 
 @RunWith(JUnitQuickcheck.class)
 public class ReservationUseCaseTests {
 
-  private LocalDate aDay = LocalDate.of(2021,6,4);
+  private LocalDate aDay = LocalDate.now();
 
   private Reservation fromRequestedCapacity(int requested) {
     return new Reservation(aDay,"happy customer", "happy@customer.com", requested);
@@ -46,7 +45,7 @@ public class ReservationUseCaseTests {
   @Property
   public void checkInitialCapacity() {
     ReservationUseCase uc = new ReservationUseCase(kvs, log);
-    assertEquals(uc.maxCapacity(), uc.availableSeats(aDay));
+    assertEquals(uc.maxCapacity(), uc.availableSeats(aDay,uc.getKvs(),uc.getTrace()));
   }
 
 }
